[project @ 2004-08-15 20:28:02 by panne] <fanfare>Finally: Support for DocBook XML!</fanfare>
After endless frustrating hours, I came to the conclusion that using (Open)Jade
for DocBook XML is virtually impossible, there are *tons* of problems with it:
Wildly differing paths, incorrect stylesheets, broken catalogs at funny places,
broken wrapper scripts, missing support on Cygwin etc.  >:-P * * *

So we follow the ubiquitous XSL hype and use xsltproc + DocBook XSL stylesheets
to transform DocBook XML to HTML and FO. From FO there are two routes to PDF and
PostScript: Either via FOP (preferred) or via PassiveTeX. Validation can be done
via xmllint using the new make target "validate". When PassiveTeX is available,
DVI can be generated, too.  A new make target "no-chunks-html" is available for
generating an all-in-one HTML document. Currently there is no way to generate
plain text or RTF, but at least the former could easily be added.

Generating HTML works out-of-the-box on Cygwin now, but you will have to install
FOP for other formats, because there seems to be no standard package for it.

The HTML appearance can be controlled via CSS, but the current location of the
master stylesheet below fptools/mk is a bit debatable. Better suggestions are
welcome.

Currently there is still support for SGML documents, but it will be dropped when
all documents are converted to DocBook XML. The build system is complex enough
already with the support for a *single* kind of DocBook...