[project @ 2004-11-11 16:07:14 by simonmar] The new ghc-pkg tool.
So far, the new syntax is supported, and the format of the package
definitions has changed to InstalledPackageInfo rather than the old
PackageConfig type.  The format of the package.conf file is now
[InstalledPackageInfo] (using show/read).

We still support the old ghc-pkg command line syntax, and the old
PackageConfig syntax for package definitions.  These are deprecated
features, of course.

Not much is done with the list of exposed/hidden modules, or versions
(yet).