[project @ 2004-11-29 16:25:03 by simonpj] --------------------- 	Simplify ImportAvails 	---------------------
Every Name has, for some while, contained its "parent";
the type or class inside which it is defined.  But the rest
of the renamer wasn't using this information as much as it 
could do.  In particular, the ImportAvails type was more elaborate
than necessary.

This commit combines these two fields of ImportAvails:
	imp_env :: AvailEnv
	imp_qual :: ModuleEnv AvailEnv
into one
	imp_env :: ModuleEnv NameSet 

This is quite a bit simpler.  Less redundancy and, I think, less
code.