[project @ 2004-12-22 12:04:14 by simonpj] -------------------------- 	Add -fwarn-orphans flag 	--------------------------
This gives a decent report for modules that contain 'orphan' instance and
rule declarations.  These are to be avoided, because GHC has to proactively
read the interface file every single time, just in case the instance/rule is
needed.

The flag just gives a convenient way of identifying the culprits.