[project @ 2005-01-14 17:57:41 by simonmar] HEADS UP!  You now need to use an up to date Happy from CVS to build GHC.  Happy version 1.15 will be released shortly.
Replace the slow hacked up String-based GetImports with one based on
the real Haskell parser.  This requires a new addition to Happy to
support parsing partial files.  We now avoid reading each source file
off the disk twice: once to get its module name and imports, and again
to parse it.  Instead we just slurp it once, and cache the StringBuffer.

This should result in improved startup times for ghc --make,
especially when there are lots of source files.