[project @ 2005-01-28 17:44:55 by simonpj] Arrange that when seeking instance decls in GHCi, in response to a :info command, we only print ones whose types are in scope unqualified.  This eliminates an alarmingly long list when simply typing ':info Show', say.
On the way, I reorganised a bit.  GHCi printing happens by
converting a TyThing to an IfaceDecl, and printing that.
I now arrange to generate unqualifed IfaceExtNames directly
during this conversion, based on what is in scope.  Previously
it was done during the pretty-printing part via the UserStyle.
But this is nicer.