[project @ 2005-02-07 12:07:21 by simonpj] ------------------------------ 	Improve unused-import warnings 	------------------------------
	Merge to STABLE

This commit improves the warning messages for unused imports, in the
case where the 'module A' form is used in an export list.

In doing this, I've realised that the unused-import checking is deficient
in several ways.  At least

	a) it doesn't recognise that there might be several import statements
	   for the same module (TcRnTypes.imp_mods has only one entry per
	   module

	b) it doesn't understand about qualified modules at all

	c) even more fundamentally, it starts from the used Names,
	   but if the module mentions (say) aliases M.f and N.f for
	   the same Name, then two imports might be necessary for it

I'm not going to fix these problems now; this message just records them.