[project @ 2005-02-07 12:16:50 by simonpj] ----------------------------------------------------- 	Switch off derivable type class generation by default 	-----------------------------------------------------
	Merge to STABLE

This commit switches off the generation of the support code for 'Derivable
Type Classes' by default.  Use -fgenerics to switch it on.

Motivation: seldom used, and there's a nasty blowup in Core types for data
types that have lots of constructors or lots of fields.

We now put -fgenerics in the GhcLibOpts, so that the libraries have the
support code, as before.