[haddock @ 2005-02-28 16:22:08 by simonmar] Attempt to fix the layout of the package names in the contents.
Having tried just about everything, the only thing I can get to work
reliably is to make the package names line up on a fixed offset from
the left margin.  This obviously isn't ideal, so anyone else that
would like to have a go at improving it is welcome.  One option is to
remove the +/- buttons from the contents list and go back to a plain
table.

The contents page now uses CSS for layout rather than tables.  It
seems that most browsers have different interpretations of CSS layout,
so only the simplest things lead to consistent results.