[project @ 2005-04-04 13:51:26 by simonmar] Big cleanup of the scheduler.
The main idea here was to extract as much stuff as possible from the
scheduler loop into seprate functions, so as to better expose the
control structure of the scheduler.  Now, the scheduler loop is down
to some 300+ lines; there's some more code that could be extracted,
but I think it looks pretty good now.

This work is partly due to an initial cleanup by the GRAN/PAR folks,
heavily re-worked by me.