[project @ 2005-04-25 13:02:31 by simonmar] Requested by Duncan Coutts: -optc flags get passed to the assembler phase too.  It's not obvious, when using GHC to compile a .c file, that options need to be passed both with -optc and -opta - gcc doesn't need this, for example.
This is an experimental change - if it breaks anything we can back it
out.  If anyone needs to pass flags to the .c->.s phase *only*, then
we will need to add another flag for that purpose.  It seems unlikely,
though (there aren't many flags that affect the assembler only, and
those that do don't conflict with flags for the C phase).